requirements.txt
pycryptodome>=3.9.9
fpdf>=1.7.2
